refactoring UserBusinessService class
